Move CachedTransform to OpenRA.Primitives.
