+++ OpenRA.Mods.Common/Traits/World/TimeLimitManager.cs
@@ -12,6 +12,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using OpenRA.Mods.Common.Widgets;
+using OpenRA.Primitives;
 using OpenRA.Traits;
 using OpenRA.Widgets;
