+++ OpenRA.Mods.Common/Widgets/Logic/Ingame/DebugLogic.cs
@@ -11,6 +11,7 @@
 
 using OpenRA.Graphics;
 using OpenRA.Mods.Common.Traits;
+using OpenRA.Primitives;
 using OpenRA.Widgets;
 
 namespace OpenRA.Mods.Common.Widgets.Logic.Ingame
