+++ OpenRA.Mods.Common/Widgets/Logic/Ingame/DebugMenuLogic.cs
@@ -11,6 +11,7 @@
 
 using System;
 using OpenRA.Mods.Common.Traits;
+using OpenRA.Primitives;
 using OpenRA.Traits;
 using OpenRA.Widgets;
