+++ OpenRA.Mods.Common/Widgets/Logic/Ingame/GameTimerLogic.cs
@@ -12,6 +12,7 @@
 using System;
 using OpenRA.Mods.Common.Traits;
 using OpenRA.Network;
+using OpenRA.Primitives;
 using OpenRA.Widgets;
 
 namespace OpenRA.Mods.Common.Widgets.Logic
