+++ OpenRA.Mods.Common/Widgets/Logic/Ingame/IngameChatLogic.cs
@@ -15,6 +15,7 @@ using OpenRA.Mods.Common.Commands;
 using OpenRA.Mods.Common.Lint;
 using OpenRA.Mods.Common.Traits;
 using OpenRA.Network;
+using OpenRA.Primitives;
 using OpenRA.Widgets;
 
 namespace OpenRA.Mods.Common.Widgets.Logic
