+++ OpenRA.Mods.Common/Widgets/Logic/Settings/AudioSettingsLogic.cs
@@ -13,6 +13,7 @@ using System;
 using System.Linq;
 using OpenRA.Graphics;
 using OpenRA.Mods.Common.Traits;
+using OpenRA.Primitives;
 using OpenRA.Widgets;
 
 namespace OpenRA.Mods.Common.Widgets.Logic
